requirements.txt
jsonrpclib-pelix>=0.4.3.2
PyYAML>=6.0
colorama
prettytable
paramiko
scp
tqdm
requests
cvprac>=1.2.0
netaddr>=0.8.0
pydantic>=1.9.1
Jinja2>=3.1.2
tabulate>=0.8.10
rich>=12.5.1
